requirements.txt
google-generativeai
python-dotenv
streamlit
pillow
pdf2image
sqlite3
